# apps/server/openapi.json
{
  "components": {
    "schemas": {
      "AlertHistory": {
        "description": "Alert delivery history record (audit log and retry queue)",
        "properties": {
          "alert_rule_id": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "alert_type": {
            "type": "string"
          },
          "attempt_count": {
            "format": "int32",
            "type": "integer"
          },
          "channel_id": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "channel_name": {
            "type": "string"
          },
          "channel_type": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "error_message": {
            "nullable": true,
            "type": "string"
          },
          "http_status_code": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "idempotency_key": {
            "type": "string"
          },
          "issue_id": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "next_retry_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "project_id": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "sent_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus"
          }
        },
        "required": [
          "id",
          "alert_type",
          "channel_type",
          "channel_name",
          "status",
          "attempt_count",
          "idempotency_key",
          "created_at"
        ],
        "type": "object"
      },
...
      "AlertStatus": {
        "description": "Status of an alert delivery attempt",
        "enum": [
          "pending",
          "sent",
          "failed",
          "skipped"
        ],
        "type": "string"
      },
...
      "ChannelType": {
        "description": "Type of notification channel",
        "enum": [
          "webhook",
          "email",
          "slack"
        ],
        "type": "string"
      },
      "CreateAlertRule": {
        "description": "DTO for creating an alert rule",
        "properties": {
          "alert_type": {
            "$ref": "#/components/schemas/AlertType"
          },
          "channel_ids": {
            "items": {
              "format": "int32",
              "type": "integer"
            },
            "type": "array"
          },
          "conditions": {},
          "cooldown_minutes": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "alert_type"
        ],
        "type": "object"
      },
...
      "CreateNotificationChannel": {
        "description": "DTO for creating a notification channel",
        "properties": {
          "channel_type": {
            "$ref": "#/components/schemas/ChannelType"
          },
          "config": {},
          "is_enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "channel_type",
          "config"
        ],
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "description": "JSON error response structure",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "NotificationChannel": {
        "description": "Global notification channel (e.g., Slack workspace, webhook endpoint)",
        "properties": {
          "channel_type": {
            "$ref": "#/components/schemas/ChannelType"
          },
          "config": {},
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "failure_count": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "is_enabled": {
            "type": "boolean"
          },
          "last_failure_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "last_failure_message": {
            "nullable": true,
            "type": "string"
          },
          "last_success_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "channel_type",
          "config",
          "is_enabled",
          "failure_count",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "TestChannelResponse": {
        "description": "Response for test channel endpoint",
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "type": "object"
      },
...
      "UpdateNotificationChannel": {
        "description": "DTO for updating a notification channel",
        "properties": {
          "config": {},
          "is_enabled": {
            "nullable": true,
            "type": "boolean"
          },
          "name": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "bearerFormat": "hex",
        "description": "40-character hex API token",
        "scheme": "bearer",
        "type": "http"
      },
      "sentry_auth": {
        "description": "Sentry SDK authentication header or ?sentry_key= query parameter",
        "in": "header",
        "name": "X-Sentry-Auth",
        "type": "apiKey"
      },
      "session_auth": {
        "in": "cookie",
        "name": "rustrak_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Rustrak Team"
    },
    "description": "Ultra-lightweight error tracking server compatible with Sentry SDKs",
    "license": {
      "name": "GPL-3.0"
    },
    "title": "Rustrak API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/alert-channels": {
      "get": {
        "operationId": "list_channels",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationChannel"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of notification channels"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/alert-channels",
        "tags": [
          "alerts"
        ]
      },
      "post": {
        "operationId": "create_channel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationChannel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationChannel"
                }
              }
            },
            "description": "Channel created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "POST /api/alert-channels",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alert-channels/{id}": {
      "delete": {
        "operationId": "delete_channel",
        "parameters": [
          {
            "description": "Channel ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Channel deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Channel not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "DELETE /api/alert-channels/{id}",
        "tags": [
          "alerts"
        ]
      },
      "get": {
        "operationId": "get_channel",
        "parameters": [
          {
            "description": "Channel ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationChannel"
                }
              }
            },
            "description": "Channel details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Channel not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/alert-channels/{id}",
        "tags": [
          "alerts"
        ]
      },
      "patch": {
        "operationId": "update_channel",
        "parameters": [
          {
            "description": "Channel ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationChannel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationChannel"
                }
              }
            },
            "description": "Channel updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Channel not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "PATCH /api/alert-channels/{id}",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alert-channels/{id}/test": {
      "post": {
        "operationId": "test_channel",
        "parameters": [
          {
            "description": "Channel ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestChannelResponse"
                }
              }
            },
            "description": "Test result"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Channel not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "POST /api/alert-channels/{id}/test",
        "tags": [
          "alerts"
        ]
      }
    },
...
    "/api/projects/{project_id}/alert-history": {
      "get": {
        "operationId": "list_history",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertHistory"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Alert history"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/projects/{project_id}/alert-history",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/projects/{project_id}/alert-rules": {
      "get": {
        "operationId": "list_rules",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertRuleResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of alert rules"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/projects/{project_id}/alert-rules",
        "tags": [
          "alerts"
        ]
      },
      "post": {
        "operationId": "create_rule",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRule"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleResponse"
                }
              }
            },
            "description": "Rule created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "POST /api/projects/{project_id}/alert-rules",
        "tags": [
          "alerts"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer_auth": []
    },
    {
      "session_auth": []
    },
    {
      "sentry_auth": []
    }
  ],
...
}